Move getSqlQueryForIndexCreateOrEdit()
