+++ psalm-baseline.xml
@@ -11466,6 +11466,14 @@
       <code><![CDATA[$columnMeta['DefaultValue']]]></code>
     </PossiblyUndefinedArrayOffset>
   </file>
+  <file src="src/Table/Indexes.php">
+    <PossiblyInvalidArgument>
+      <code>$oldIndex</code>
+    </PossiblyInvalidArgument>
+    <PossiblyNullOperand>
+      <code><![CDATA[$column->getSubPart()]]></code>
+    </PossiblyNullOperand>
+  </file>
   <file src="src/Table/Search.php">
     <MixedArgument>
       <code><![CDATA[$_POST['criteriaColumnNames'][$columnIndex]]]></code>
@@ -11736,7 +11744,6 @@
     <PossiblyInvalidArgument>
       <code><![CDATA[$_POST['constraint_name'][$masterFieldMd5]]]></code>
       <code><![CDATA[$fields[$column]['expr']]]></code>
-      <code>$oldIndex</code>
       <code><![CDATA[$options['expr']]]></code>
     </PossiblyInvalidArgument>
     <PossiblyInvalidArrayOffset>
@@ -11776,7 +11783,6 @@
       <code><![CDATA[$statement->altered]]></code>
     </PossiblyNullIterator>
     <PossiblyNullOperand>
-      <code><![CDATA[$column->getSubPart()]]></code>
       <code><![CDATA[Generator::mysqlDie($tmpErrorCreate, $createQuery, false, '', false)]]></code>
       <code><![CDATA[Generator::mysqlDie($tmpErrorDrop, $dropQuery, false, '', false)]]></code>
     </PossiblyNullOperand>
@@ -15270,19 +15276,6 @@
       <code>providerColumnMetaDefault</code>
     </PossiblyUnusedMethod>
   </file>
-  <file src="tests/classes/Table/IndexesTest.php">
-    <DeprecatedMethod>
-      <code>Config::getInstance()</code>
-      <code>DatabaseInterface::getInstance()</code>
-    </DeprecatedMethod>
-    <MixedMethodCall>
-      <code>method</code>
-      <code>willReturn</code>
-    </MixedMethodCall>
-    <UndefinedMethod>
-      <code>expects</code>
-    </UndefinedMethod>
-  </file>
   <file src="tests/classes/Table/Maintenance/MessageTest.php">
     <PossiblyUnusedMethod>
       <code>providerForTestFromArray</code>
@@ -15297,8 +15290,6 @@
     <DeprecatedMethod>
       <code>Config::getInstance()</code>
       <code>Config::getInstance()</code>
-      <code>DatabaseInterface::getInstance()</code>
-      <code>DatabaseInterface::getInstance()</code>
       <code>DatabaseInterface::getInstance()</code>
       <code>DatabaseInterface::getInstance()</code>
       <code>DatabaseInterface::getInstance()</code>
