+++ src/Table/Table.php
@@ -1876,120 +1876,6 @@ class Table implements Stringable
     }
 
     /**
-     * Function to get the sql query for index creation or edit
-     *
-     * @param Index         $index current index
-     * @param Message|false $error whether error occurred or not
-     */
-    public function getSqlQueryForIndexCreateOrEdit(Index $index, Message|false &$error): string
-    {
-        // $sql_query is the one displayed in the query box
-        $sqlQuery = sprintf(
-            'ALTER TABLE %s.%s',
-            Util::backquote($this->dbName),
-            Util::backquote($this->name),
-        );
-
-        // Drops the old index
-        if (isset($_POST['old_index'])) {
-            $oldIndex = is_array($_POST['old_index']) ? $_POST['old_index']['Key_name'] : $_POST['old_index'];
-            if ($oldIndex === 'PRIMARY') {
-                $sqlQuery .= ' DROP PRIMARY KEY,';
-            } else {
-                $sqlQuery .= sprintf(
-                    ' DROP INDEX %s,',
-                    Util::backquote($oldIndex),
-                );
-            }
-        }
-
-        // Builds the new one
-        switch ($index->getChoice()) {
-            case 'PRIMARY':
-                if ($index->getName() == '') {
-                    $index->setName('PRIMARY');
-                } elseif ($index->getName() !== 'PRIMARY') {
-                    $error = Message::error(
-                        __('The name of the primary key must be "PRIMARY"!'),
-                    );
-                }
-
-                $sqlQuery .= ' ADD PRIMARY KEY';
-                break;
-            case 'FULLTEXT':
-            case 'UNIQUE':
-            case 'INDEX':
-            case 'SPATIAL':
-                if ($index->getName() === 'PRIMARY') {
-                    $error = Message::error(
-                        __('Can\'t rename index to PRIMARY!'),
-                    );
-                }
-
-                $sqlQuery .= sprintf(
-                    ' ADD %s',
-                    $index->getChoice(),
-                );
-                if ($index->getName() !== '') {
-                    $sqlQuery .= ' ' . Util::backquote($index->getName());
-                }
-
-                break;
-        }
-
-        $indexFields = [];
-        foreach ($index->getColumns() as $key => $column) {
-            $indexFields[$key] = Util::backquote($column->getName());
-            if (! $column->getSubPart()) {
-                continue;
-            }
-
-            $indexFields[$key] .= '(' . $column->getSubPart() . ')';
-        }
-
-        if ($indexFields === []) {
-            $error = Message::error(__('No index parts defined!'));
-        } else {
-            $sqlQuery .= ' (' . implode(', ', $indexFields) . ')';
-        }
-
-        $keyBlockSizes = $index->getKeyBlockSize();
-        if ($keyBlockSizes !== 0) {
-            $sqlQuery .= ' KEY_BLOCK_SIZE = ' . $keyBlockSizes;
-        }
-
-        // specifying index type is allowed only for primary, unique and index only
-        // TokuDB is using Fractal Tree, Using Type is not useless
-        // Ref: https://mariadb.com/kb/en/storage-engine-index-types/
-        $type = $index->getType();
-        if (
-            $index->getChoice() !== 'SPATIAL'
-            && $index->getChoice() !== 'FULLTEXT'
-            && in_array($type, Index::getIndexTypes(), true)
-            && ! $this->isEngine('TOKUDB')
-        ) {
-            $sqlQuery .= ' USING ' . $type;
-        }
-
-        $parser = $index->getParser();
-        if ($index->getChoice() === 'FULLTEXT' && $parser !== '') {
-            $sqlQuery .= ' WITH PARSER ' . $parser;
-        }
-
-        $comment = $index->getComment();
-        if ($comment !== '') {
-            $sqlQuery .= sprintf(
-                ' COMMENT %s',
-                $this->dbi->quoteString($comment),
-            );
-        }
-
-        $sqlQuery .= ';';
-
-        return $sqlQuery;
-    }
-
-    /**
      * Function to handle update for display field
      *
      * @param string $displayField display field
