+++ tests/classes/Table/IndexesTest.php
@@ -4,67 +4,52 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\Tests\Table;
 
-use PhpMyAdmin\Config;
 use PhpMyAdmin\Current;
 use PhpMyAdmin\DatabaseInterface;
 use PhpMyAdmin\Index;
+use PhpMyAdmin\Message;
 use PhpMyAdmin\Table\Indexes;
 use PhpMyAdmin\Table\Table;
 use PhpMyAdmin\Tests\AbstractTestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\MockObject\MockObject;
 
 #[CoversClass(Indexes::class)]
 class IndexesTest extends AbstractTestCase
 {
+    private DatabaseInterface&MockObject $dbi;
+
     protected function setUp(): void
     {
         parent::setUp();
 
-        /**
-         * SET these to avoid undefined index error
-         */
         Current::$database = 'db';
         Current::$table = 'table';
-        $config = Config::getInstance();
-        $config->selectedServer['pmadb'] = '';
-        $config->selectedServer['DisableIS'] = false;
-        $GLOBALS['urlParams'] = ['db' => 'db', 'server' => 1];
 
-        $dbi = $this->getMockBuilder(DatabaseInterface::class)
+        $this->dbi = $this->getMockBuilder(DatabaseInterface::class)
             ->disableOriginalConstructor()
             ->getMock();
-
-        $indexs = [
-            ['Schema' => 'Schema1', 'Key_name' => 'Key_name1', 'Column_name' => 'Column_name1'],
-            ['Schema' => 'Schema2', 'Key_name' => 'Key_name2', 'Column_name' => 'Column_name2'],
-            ['Schema' => 'Schema3', 'Key_name' => 'Key_name3', 'Column_name' => 'Column_name3'],
-        ];
-
-        $dbi->expects($this->any())->method('getTableIndexes')
-            ->willReturn($indexs);
-
-        DatabaseInterface::$instance = $dbi;
-
-        //$_SESSION
     }
 
     public function testDoSaveData(): void
     {
-        $sqlQuery = 'ALTER TABLE `db`.`table` DROP PRIMARY KEY, ADD UNIQUE ;';
+        $sqlQuery = 'ALTER TABLE `db`.`table` DROP PRIMARY KEY, ADD PRIMARY KEY (`id`);';
 
         $table = $this->getMockBuilder(Table::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $table->expects($this->any())->method('getSqlQueryForIndexCreateOrEdit')
-            ->willReturn($sqlQuery);
-
-        $dbi = DatabaseInterface::getInstance();
-        $dbi->expects($this->any())->method('getTable')
+        $this->dbi->expects($this->any())->method('getTable')
             ->willReturn($table);
 
         $index = new Index();
+        $index->set([
+            'Key_name' => 'PRIMARY',
+            'columns' => [['Column_name' => 'id']],
+        ]);
 
-        $indexes = new Indexes($dbi);
+        $indexes = new Indexes($this->dbi);
+
+        $_POST['old_index'] = 'PRIMARY';
 
         // Preview SQL
         $sqlResult = $indexes->doSaveData($index, false, Current::$database, Current::$table, true);
@@ -77,9 +62,31 @@ class IndexesTest extends AbstractTestCase
         $this->assertStringContainsString($sqlQuery, $sqlResult);
 
         // Error message
-        // Cannot be tested at the moment.
-        // $index->setName('PRIMARY'); // Cannot rename any index to primary so the operation should fail
-        // $indexes->doSaveData($index, false, Current::$database, Current::$table, false);
-        // $this->assertInstanceOf(Message::class, $sqlResult);
+        $index->setName('NOT PRIMARY'); // Cannot rename primary so the operation should fail
+        $sqlResult = $indexes->doSaveData($index, false, Current::$database, Current::$table, false);
+        $this->assertInstanceOf(Message::class, $sqlResult);
+    }
+
+    public function testGetSqlQueryForIndexCreateOrEdit(): void
+    {
+        $table = $this->getMockBuilder(Table::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+        $this->dbi->expects($this->any())->method('getTable')
+            ->willReturn($table);
+        $indexes = new Indexes($this->dbi);
+
+        $db = 'pma_db';
+        $table = 'pma_table';
+        $index = new Index();
+
+        $_POST['old_index'] = 'PRIMARY';
+        $sql = $indexes->getSqlQueryForIndexCreateOrEdit($db, $table, $index);
+        $this->assertEquals('ALTER TABLE `pma_db`.`pma_table` DROP PRIMARY KEY, ADD UNIQUE;', $sql);
+
+        $_POST['old_index'] = [];
+        $_POST['old_index']['Key_name'] = 'PRIMARY';
+        $sql = $indexes->getSqlQueryForIndexCreateOrEdit($db, $table, $index);
+        $this->assertEquals('ALTER TABLE `pma_db`.`pma_table` DROP PRIMARY KEY, ADD UNIQUE;', $sql);
     }
 }
